fix(restaking): change to `[]byte(addr.String())`
The prefix iterator requires a []byte key, and `addr.String` is not a
byte slice (obviously). However, using `addr.Bytes` results in the raw
bytes being used, which is inconsistent with `addr.String` which is
bech32 encoded. Since the rest of the module uses `addr.String` as the
key, we follow the same process here.