feat(dogfood): add parameters
The dogfooding module, as previously described, is used as the staking
module for the Exocore chain. It allows operators and delegators to
deposit and delegate their assets on a client chain for staking. The
staked assets are then used to secure the Exocore chain.

The parameters used by this module are the following:
 - EpochsUntilUnbonded, which represents the number of epochs after
   which an unbonding is effective.
 - EpochIdentifier, which is the identifier of the epoch duration. It
   should be valid according to the epoch keeper of `app.go`, and is
   thus constrained to the options week/day/hour.
 - MaxValidators, which is the maximum number of validators that will be
   forwarded to ABCI.
 - HistoricalEntries, which is the number of historical entries to
   persist in state. These are used by IBC.