feat(dogfood): load genesis state for bootstrap
The bootstrapping smart contract is responsible for accepting user
deposits, operator registration, and delegations. These operations may
flow freely until 24 hours (say) before the spawn time of Exocore, at
which the contract will be locked. During this time, the genesis state
of Exocore will be amended to record the initial deposits, delegations,
operator registrations and public keys. The last of these will feed into
the dogfooding module which will mark these operators as part of the
initial validator set and allow block production.

This PR includes the capability to load that state, and save the newly
created operator information to disk. In addition, the staking hooks
have been set up (partially) in this PR, which allow the SDK's slashing
keeper to record validator signing rates for downtime slashing.