docs(dogfood): explain the operator queue logic
The operator module should restrict an operator from doing things that
are logically not possible. For example, an operator that is already
opted in should not be able to opt in again. Similarly, an operator who
has opted out of a chain should not be able to replace the key for that
chain without opting back in. Based on these assumptions, the comment
maps out the list of possible operations that can happen within an epoch
and is used to ensure that all of the cases result in the unbonding
period being enforced accurately.