feat(dogfood): implement delegation hooks
Whenever a delegation or undelegation event happens in the delegation
module, the hooks are called. For both of these events, as long as the
operator is not in the process of opting out from the chain, consensus
key operations are queued to be applied at the end of the current epoch.

For undelegation, specifically, the `recordKey` identifier of the
undelegation is used to mark within the delegation module that the
undelegation should be held until released at some point in the future
(as opposed to releasing it after a fixed number of blocks, which is the
default behaviour). This point in the future is calculated as the
earlier of the unbonding period end, or the operator's opt out period
end (if the operator is opting out).