feat(exomint): introduce module
The exomint module is a variation of the Cosmos SDK's mint module. The
mint module depends mandatorily on the staking token supply and
optionally on the bonded ratio of the token to calculate the amount of
tokens to be minted in the block. Since none of these two metrics are
super meaningful to us in our restaking environment, we are using a
custom block reward minting module.

The module requires two parameters as input: the `MintDenom` which
defaults to the default bond denomination of the chain and the
`BlockReward` which defaults to 1. In other words, by default, this
module mints 1 stake denomination worth of coins at the beginning of the
block to the fee collector address.

The module name was chosen as `exomint` to ensure that it does not clash
with the SDK's module name (for clarity) while keeping go package naming
conventions (no underscores) in mind.