chore(lint): fix tests
a bunch of tests started failing after the switchover from evmos to
exocore. this change fixes those tests. however, do note that in some
cases the dependency of evmos v14 has resulted in duplicate code being
shared between ExocoreNetwork/exocore and ExocoreNetwork/evmos (v14).
For an example of this case, see types/errors.go which registers the
ErrInvalidChainID error for evmos as well as exocore codespace, which
should ideally not happen.