feat(dogfood): retrieve val set from id
The validator set is not bound to change at each block. It can only
change at the end of an epoch, or following a slashing event. Therefore,
storing the historical information for each block is redundant. Instead,
we create a validator set id corresponding to each validator set, and
map each height to a validator set id. The mapping and the validator set
id are pruned once the historical info for a height is not required.

TODO: actually store the validator set against each id, within the
`EndBlock` function at the end of the epoch.