feat(dogfood): implement sdk staking interface (#123)
* feat(dogfood): implement sdk staking interface

The interface required by IBC was already implemented in
`validators.go`. This PR takes that a step further and implements the
interfaces required to use the dogfood module as a drop-in replacement
for the SDK's staking module. The interfaces implemented are those
expected by the slashing, evidence and `genutil` modules. An explanation
has been provided above each function to explain why and where it is
called, and if its implementation is really necessary. There are still
some TODOs left in this PR that depend on the overall slashing /
operator design.

This branch is merged into dogfood-part6, which forms the basis of #122.

* fix(dogfood): store val set id + val set

at genesis, the val set id starts with 1. each time the validator set
changes, the val set id of the current block is retrieved. for the next
block, the val set is is stored as this id + 1.

* fix(dogfood): increment val set id correctly

in the case of genesis, we should not use height + 1 as the key in the
mapping; rather, it should be height. in all other cases, the mapping
key is height + 1. the value is the val set id in all cases.