chore(lint): fix super linter (again) (#14)
* chore(lint): satisfy gitleaks

If a variable name contains the word KEY, gitleaks does not like it and
flags it as a potential secret. I have changed the name of the variable
in these files from KEYALGO to ALGO since it is not a secret but rather
the algorithm of the key generation. That silences gitleaks.

* chore(lint): placate YAML linter

added a blank line at the end of `.golangci.yml`

* chore(lint): activate Docker / IAAS linter

* chore(lint): placate shellcheck

* chore(lint): placate markdownlint

* chore(lint): placate markdownlint from `lint.yml`

The previous commit worked on the markdownlint from the super-linter.

* build: use test keyring in localnet

The OS keyring may not always be available.

* chore(lint): merge docker RUN statements

It can potentially reduce the number of layers in the image, making it
smaller and more efficient to build and pull.

* fix(localnet): remove seed URL upon init

* chore(lint): some lint suggestions for Dockerfile

Note that this file is used from the root of the project via `make
localnet-build`, which is why the paths work.

* chore(lint): apply shfmt

* chore(lint): add docker healthcheck and user

* chore(build): update release go version to 1.21

* chore(ci): add permissions to workflows

* chore(ci): skip generated files

* chore(ci): placate yaml linter

* chore(ci): use correct argument name for checkov

* chore(ci): add top level permissions to codeql

* chore(ci): add json super linter to match develop

* chore(ci): disable JSON super linter

* chore(lint): trigger lint workflows with comment

* chore(deps): upgrade protobuf to 1.33.0

CVE-2024-24786 / GO-2024-2611 is fixed with this upgrade. Even though
this package is fundamental to our system, the changes between 1.31.0
and 1.33.0 are minor.

* Revert "chore(deps): upgrade protobuf to 1.33.0"

This reverts commit c09f9864441e9aaf64ad2d31bb568eb71bf6eb35.

* chore(deps): update protobuf

A vulnerability has been fixed in v1.33.0 of google.golang.org/protobuf
and we must update the associated github.com/golang/protobuf as well.

* chore(ci): coverage badge comment on PR

* chore(build): exclude testutil from coverage

* chore(ci): add test coverage commenting workflow

* chore(ci): use step output instead of env

* chore(ci): add commit hash in test comment

* chore(ci): remove deprecated output

The test workflow will not fully run right now because it runs on
`pull_request_target` which must be merged into the base branch first.