fix(ci): semgrep, cov-comment, buf-break-no-fail (#15)
* chore(lint): security options in docker compose

* chore(lint): unlock mutex in testutil

* chore(lint): limit value of uint using strconv

* fix(lint): check out old code correctly for diff

In the event of a pull_request_target action, the checkout action by
default does not check out the PR but the base repository. This change
explicitly sets it up to check out the PR's head branch.

* Revert "fix(lint): check out old code correctly"

This reverts commit 4f24ba223548593149602c2bb8842c38fab2dd13.

* chore(lint): split test coverage commenter

See comment on test.yml for explanation.

* chore(ci): trigger go based CIs

delete superfluous comment

* fix(lint): add empty line at end of yaml file

* chore(lint): fix the test-comment workflow

- shell check the script
- add top level permissions

* Revert "chore(lint): unlock mutex in testutil"

This reverts commit a6cdb02745b6afe7d7c60277895921e5654d9631. It is said
that the unlocking of the mutex might deviate from the original purpose
of the function. Instead, we are electing to ignore the testutil folder,
which will be sent in the next commit.

* chore(ci): ignore testutil for semgrep

* chore(ci): do not fail if breaking proto changes

* chore(ci): trigger buf-breaking-action check

This commit will be reverted before merge of the PR.

* chore(ci): add comment if buf reports change

* fix(ci): only overwrite prev coverage comment

* fix(ci)(proto): +comment read/write permission

* fix(ci): add comment rw permission to job

* fix(ci): update job name in test-comment

* fix(ci): give `proto.yml` all rw permissions

* fix(ci): split buf-breaking and its comment

As before, the `pull_request` event does not have enough permissions to
make a comment and the `pull_request_target` cannot safely check out the
PR. Hence, we split the commenting process into 2 parts:
 - Run the buf breaking action and do not fail the workflow even if it
   errors
 - In a separate workflow, consume the artifact produced by the previous
   workflow and comment (or append to an existing comment) its contents.
Note that the *-comment.yml workflows are being commented blindly at the
moment because it is not possible to run them until they exist on the
branch.

* fix(ci): if condition shouldn't be always true

* Revert "chore(ci): trigger go based CIs"

This reverts commit 62b3744cc65d994d28c0c9ef69c06e0d4239bb8e.

* Revert "chore(ci): trigger buf-breaking-action...

check." This reverts commit 88566fdfb4ed74935374c4062b950f35beaf8e4b.