fix(delegation): set up genesis file loading
The delegation module by itself does not store a lot of information. The
small portion that it does store is handled by this commit at the time
of genesis. It validates that the formats of the StakerID, the AssetID
and the OperatorAddress are correct and that the amount is not nil.