fix(app): clarify claims and recovery modules
The pre-existing claims and recovery modules are used for airdrops and
token recovery (between Cosmos and Ethereum addresses) over authorized
IBC channels, respectively. This commit explicitly disables the airdrop
feature while retaining to claims module to enable recovery of tokens.
Note that the authorized channels for the recovery are still undecided.