fix(delegation): do not use genesis map
Since Cosmos uses an IAVL+ based tree, the order of (unrelated) edits to
the leaf nodes can impact the state root. Even though protobuf
guarantees the order of a map when serializing it, it cannot guarantee
that the order will be maintained once the map is deserialized into
Golang. Hence, we cannot use a map during genesis (or anywhere
meaningful at all).