refactor(dogfood,operator): source module types
Since the interfaces of events are now well defined in the corresponding
modules, we can move our interface guards to validate them. For example,
the implementation of delegation hooks in the dogfood module was guarded
by `types.DelegationHooks`, which has now been changed to
`delegationtypes.DelegationHooks`. This ensures the full compatibility
of the interface from the get-go.