refactor(dogfood): store full header not subset
Unfortunately, our prior work around of storing a subset of the header
fields for HistoricalInfo can only work for the IBC usage. The EVM
module requires the full header to be stored so that it can calculate
its hash and that of older blocks to respond to the `blockhash` calls
within Solidity.