feat(precompile): add `getClientChains` function
This function can be used by the `ExocoreGateway` to fetch all currently
registered client chains. This list is used to send a `markBootstrapped`
transaction to all the destination chains.

TODO: at this point, the function to mark bootstrap on all chains must
be triggered manually. This call should be automated in the future.