fix(build): use `test` backend for localnet
The `os` backend is not recommended for use in headless environments,
and hence we replace it with the `test` backend. Note that it stores the
keys in plaintext and is therefore simply the `file` backend without the
encryption.

For deployments of public networks in headless environments, Cosmos
recommends using `file` or `pass` backends.