fix(dogfood,operator): clear prev keys correctly
AVSs are free to choose their own epoch schedule. The previous key needs
to be cleared only at epoch end (and not at block end as it was done
previously). Additionally, it is specific to the AVS identifier (the
chainID in the case of the dogfood module) that should be cleared.

BREAKING: the format of storing the previous consensus keys has changed.
Previously, they were indexed by operator's AccAddress and chainID. Now
they are indexed by the reverse of it, that is, chainID + AccAddress.
This is done to ensure easy clearing of the data for each specific AVS.

BREAKING: the `prefixSlashAssetsState` and another prefix were sharing
the same value of 5 `BytePrefixForOperatorAndChainIDToConsKey`. This has
been fixed.