feat: CLIs for delegation, operator, dogfood (#46)
* refactor(operator): use flags to RegisterOperator

Instead of relying on the positioning of arguments, the flags mechanism
(as seen in the Cosmos SDK's x/staking) offers a cleaner approach.

* feat(operator): add commission rate to CLI

* refactor(operator): validate registration input

* feat(operator): add `OptIn` to AVS function

* refactor(assets): remove app chains

These should be in the AVS module anyway.

* fix(operator): share `opt.go`, `consensus_keys.go`

An operator must first opt into the chain-as-an-AVS before assigning a
consensus key for said chain. Similarly, to opt out, the operator must
first remove their key and then opt out.

* fix(operator): correct interface

* fix(dogfood): remove unused code

* feat(operator): add AVS opt out CLI

* fix(operator): implement msg server

* feat(operator): implement cons key interfaces

* fix(operator): skip opting out operators from list

* fix typos

* fix(dogfood): update operator hook assumptions

* feat(operator): add QueryAllOperatorKeysByChainID

* feat(operator): implement QueryOperatorConsAddress

* feat(operator): implement QueryAllOperatorConsAddr

* feat(dogfood): implement QueryOptOutsToFinish

* feat(dogfood): QueryOperatorOptOutFinishEpoch

* feat(delegation): query undelegation hold count

* feat(dogfood): implement QueryUndelegationsToMature

* fix(delegation): use string record key type

* feat(dogfood): query undelegation maturity epoch

* feat(dogfood): query validator

* fix(operator): allow build

* fix(dogfood): actually schedule cons addr prune

* chore: lint and tests

* fix: use consistent naming of query members

* remove unused comments

* doc: update comment for why chain not chain_id

* fix(operator): add Router for tx

* chore(lint)

* fix(test): fix operator tests, disable 1

* fix(proto): prevent complains about duplicate reg

* fix(operator): query operator CLI with address

* feat(operator): add set-cons-key and removal cli

* fix(dogfood,operator): clear prev keys correctly

AVSs are free to choose their own epoch schedule. The previous key needs
to be cleared only at epoch end (and not at block end as it was done
previously). Additionally, it is specific to the AVS identifier (the
chainID in the case of the dogfood module) that should be cleared.

BREAKING: the format of storing the previous consensus keys has changed.
Previously, they were indexed by operator's AccAddress and chainID. Now
they are indexed by the reverse of it, that is, chainID + AccAddress.
This is done to ensure easy clearing of the data for each specific AVS.

BREAKING: the `prefixSlashAssetsState` and another prefix were sharing
the same value of 5 `BytePrefixForOperatorAndChainIDToConsKey`. This has
been fixed.

* fix(operator): remove pointer receiver

* refactor: clarify all vs active operators

* feat(operator): add get all operators query

* fix(operator): parse cons key correctly

* chore(lint): placate proto-lint

* chore(lint): placate golang lint

* fix(operator): impl for msgs GetSignBytes, Type

Without these functions implemented, sending a transaction with the
wrong chain-id produces a panic while verifying the signature. This
panic happens on the receiving node (and not the sender), although
consensus isn't impacted.

With these functions implemented, the receiving node replies "signature
verification failed, please verify account number (%d) and chain-id
(%s)". If the transaction were only legacy amino, even the "sequence
(%d)" would have shown up in the error message.