refactor(dogfood): return cons key in `Validator`
Until `x/dogfood` was made compatible with `x/evm` and `x/oracle`, a
`nil` validator was being returned by `ValidatorByConsAddr`. This
validator was correctly processed by the `x/evidence` module when it
handled equivocation evidence.

However, with the changes introduced for compatibiltiy with `x/evm`, the
validator returned is no longer nil. Such a validator will have an
operator address, whose presence will trigger a call to `Validator`,
which is implemented by this change.