feat(dogfood): implement `ExportGenesis` (#95)
* feat(dogfood): re-scaffold genesis

In preparation for the `ExportGenesis` implementation, re-scaffold the
genesis state of the dogfood module.

* fix(dogfood): remove deprecated keys

* refactor: use new dogfood genesis

* refactor(dogfood): - validatorupdate from gen

* refactor(dogfood): update ImportGenesis

...for new genesis structure

* refactor(dogfood): simplify gen state names

* feat(dogfood): implement ExportGenesis

Partially resolves #54

* chore: lint

* proto(dogfood): remove unused import

* fix(local): compat with new dogfood genesis

* chore: lint proto file and gen

* fix(test): use correct struct member name

* chore: update go version to 1.21.11

* fix(ci): disable buggy workflows

till they can be fixed

* chore: lint

* fix(dogfood): don't stop iterating in export

The return value should be `false` to continue

* fix(doc): add comment about exported validators

* doc(dogfood): log iteration failure

If such an iteration failure happens during export, a pipe to `jq` will
fail and inform the user of the failure.

* docs(delegation): validation of record key params

* fix(app): register operator hooks for dogfood

* feat(operator): return `OptingOut` status

Whenever a query for consensus keys or addresses is answered, it should
specify whether the operator is currently in the process of opting out.
This is because the data is retained until the opt out is complete, even
though the operator may not be actively validating. The purpose of the
data retention, as always, is to slash the operator (if required), based
on the key.

* refactor(operator): rationalize opt-in/out

When an operator opts into an AVS, they should be required to supply the
consensus (or other) key to opt-in. Similarly, when they opt-out, the
key should automatically be removed.

This commit implements that functionality, however, it is restricted to
`ctx.ChainID()`. For an overbroad implementation, an AVS registry from
AVS address -> key type should be implemented.

* refactor(delegation): verbose panic

* fix(dogfood): query validator by cons addr not hex

* refactor(dogfood): move opt func to correct loc

* chore(x/operator): lint go

* fix(operator): wrap and return error

---------

Co-authored-by: cloud8little <34291844+cloud8little@users.noreply.github.com>