feat(exomint): introduce module (#96)
* feat(exomint): introduce module

The exomint module is a variation of the Cosmos SDK's mint module. The
mint module depends mandatorily on the staking token supply and
optionally on the bonded ratio of the token to calculate the amount of
tokens to be minted in the block. Since none of these two metrics are
super meaningful to us in our restaking environment, we are using a
custom block reward minting module.

The module requires three parameters as input: the `MintDenom`, the
`EpochReward` and the `EpochIdentifier`. It mints `EpochReward` coins
of `MintDenom` at the beginning of each `EpochIdentifier` by using the
hooks in the epochs module.

The module name was chosen as `exomint` to ensure that it does not clash
with the SDK's module name (for clarity) while keeping go package naming
conventions (no underscores) in mind.

* refactor(exomint): epoch minting

* fix: do not use hardcoded param

* chore: golang lint

* chore: update go version to 1.21.11

* chore: lint proto

* test(exomint): fix tests

* feat(exomint): allow update params

The function is currently guarded by the address of operator2. It should
be switched over to the governance address when it is ready.

* fix(exomint): stateful validation add

* fix(ci): ignore operator address line

* fix(exomint): add tx to cli

* feat(epochs): backport the epochs module

* fix: run go mod tidy

* fix(local): minute-wise epochs

* fix(dogfood): set default unbonding duration 7 day

* fix(local): unbond at 2 epochs

* fix(ci): fix failing jobs

* refactor(epochs): respond to ai review

* fix(docs): add `the` before module

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>

* epochs: return block time in query

The returned block time can be used to determine whether any of the
epochs are about to change.

* doc(epochs): clarify no problems with timezones

* fix(ci): provide github token to buf setup

avoid rate limits

* feat(ci): allow super-linter running with act

Example usage:
act --workflows .github/workflows/super-linter.yml pull_request

* remove unused file

* fix: typo in script

* fix(app): revert authority to gov address

The hardcoded authority is not a good design.

* doc(exomint): update comment about params

Not all params need to be supplied. Only the ones being edited. Those
that are not supplied are loaded from the previous parameter set.

* fix(exomint): return correct type from tx

* fix(exomint): do not use deprecated x/params

* test(exomint): add some tests

Currently remaining tests are ABCI test, msg test and CLI msg test

* update app for port-epochs

* fix(exomint): update for port-epochs

* feat(exomint): emit epoch id in event

* test(exomint): add epoch hook test

* fix(exomint): validate params before update

* chore(exomint): gofumpt

* fix(exomint): validate overriden params

* fix(exomint): make tests pass

* fix(exomint): `post` the UpdateParams

* fix(test): change localnet minting frequency

* fix(local): clarify comment about exomint

The default epoch identifier set in the module is `day`. The script
changes it to `minute` to enable easier localnet testing

* fix(exomint): support 0 value of rewards

In the future, a possibility that governance decides to turn off annual
inflation exists. To that end, the `exomint` module should not reject a
value of 0 for the epoch reward.

---------

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>
Co-authored-by: cloud8little <34291844+cloud8little@users.noreply.github.com>