fix(ci): always run Docker build CIs
It was wrongly assumed that this CI would only be impacted due to
changes in the Dockerfile. However, changes in Alpine Linux
repositories, for example, also cause a failrue to build.

This change also upgrades the pinned version of curl used in the
Dockerfile to one available in the repo.