fix: remove staking asset total supply
The `totalSupply` variable was created to guard against black swan
events wherein a malicious transaction could try to deposit more than
the quantity of tokens that exist. The benefits of this feature are
negligible, and the costs to maintain the constant updates between the
"real" total supply on the client chain and the "recorded" total supply
on Exocore are too high. Hence, it has been removed in its entirety.

Instead, TVL limits will be imposed by the client chain for an
operational risk management perspective.