feat(assets): remove total supply (#181)
* feat(assets): validate total supply before edit

Now that the TVL limits are imposed on the client chain instead of
Exocore, this PR renames the variable `tvlLimit` to `totalSupply`. In
addition, editing the `totalSupply` to a value lower than the currently
deposited tokens is not permitted.

* fix(doc): update IAssets comment

* feat: introduce `getTotalSupply` precompile

This can be used to fetch the total supply, according to Exocore, for an
asset.

* fix: remove staking asset total supply

The `totalSupply` variable was created to guard against black swan
events wherein a malicious transaction could try to deposit more than
the quantity of tokens that exist. The benefits of this feature are
negligible, and the costs to maintain the constant updates between the
"real" total supply on the client chain and the "recorded" total supply
on Exocore are too high. Hence, it has been removed in its entirety.

Instead, TVL limits will be imposed by the client chain for an
operational risk management perspective.

* fix the lint error caused by the lower curl version

---------

Co-authored-by: TimmyExogenous <chuang@exogenous.io>