feat(distribution): add fee-distribution and avs-reward distribution (#117)
* feat: initialize distribution

* update

* feat: add distribution logic

update

* feat: initialize fee distribution framework

* feat: add the avs reward

* feat: add the fee distribution allocation

* fix: some fixes

* fix: fix the dependency and test errors

* resolve: fix merge conflicts

resolve: fix merge conflicts

up

* chore: some refactors

* fix: fix all test errors

* chore: refactor the protobuf directory path

* some fixes

* fix: fix rebase conflict

* fix:fix consensus address parse error

* Add test dependencies and initialize genesis

* fix:ix distribution keep startup sequence

* fix:fix stored fee pool should not have been nil

* update

* feat: add reward allocation to stakers

* fix: some trivial fix

* test: add more uts

* fix: fix the operator address format

* resolve: resolve comments

* fix: fix the negative issue

* fix: fix the power update

* update

* update

* update

* fix: fix action error

* update

* fix: fix staker allocations

* update

* fix(dogfood): correct audit typos (#190)

No other recommendations were received for the x/dogfood module.

* feat(assets): remove total supply (#181)

* feat(assets): validate total supply before edit

Now that the TVL limits are imposed on the client chain instead of
Exocore, this PR renames the variable `tvlLimit` to `totalSupply`. In
addition, editing the `totalSupply` to a value lower than the currently
deposited tokens is not permitted.

* fix(doc): update IAssets comment

* feat: introduce `getTotalSupply` precompile

This can be used to fetch the total supply, according to Exocore, for an
asset.

* fix: remove staking asset total supply

The `totalSupply` variable was created to guard against black swan
events wherein a malicious transaction could try to deposit more than
the quantity of tokens that exist. The benefits of this feature are
negligible, and the costs to maintain the constant updates between the
"real" total supply on the client chain and the "recorded" total supply
on Exocore are too high. Hence, it has been removed in its entirety.

Instead, TVL limits will be imposed by the client chain for an
operational risk management perspective.

* fix the lint error caused by the lower curl version

---------

Co-authored-by: TimmyExogenous <chuang@exogenous.io>

* lint: fix github action error

* lint

* rebase

* resolve: some updates resolveing comments

* merge: merge develop

* update

* resolve: resolve further comments

* resolve: minor update

---------